more pylint cleanup
