+++ tests/conftest.py
@@ -5,6 +5,7 @@ Defines global pytest fixtures available to all tests.
 # pylint: disable=redefined-outer-name
 from pathlib import Path
 import os
+
 import pytest
