+++ tests/test_podman_compose_include.py
@@ -1,15 +1,6 @@
 from pathlib import Path
-import subprocess
 
-
-def capture(command):
-    proc = subprocess.Popen(
-        command,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-    out, err = proc.communicate()
-    return out, err, proc.returncode
+from utils import capture
 
 
 def test_podman_compose_include():
